% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.increase_month}
\alias{.increase_month}
\title{Increase a date by 1 month}
\usage{
.increase_month(d)
}
\arguments{
\item{d}{a date}
}
\value{
month + 1
}
\description{
Increase a date by 1 month
}
\keyword{internal}
