% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.recode_sinaica_units}
\alias{.recode_sinaica_units}
\title{recode units from SINAICA}
\usage{
.recode_sinaica_units(pollutant)
}
\arguments{
\item{pollutant}{type of pollutant to recode}
}
\value{
vector
}
\description{
recode units from SINAICA
}
\keyword{internal}
