% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_and_dates.R
\name{get_station_dates}
\alias{get_station_dates}
\title{Dates supported by a station}
\usage{
get_station_dates(station_id, type = c("C", "V", "M"))
}
\arguments{
\item{station_id}{numeric code of the station}

\item{type}{The type of data to download. One of the following:
\itemize{
\item{"C"}{ - Crude data that has not been validated}
\item{"V"}{ - Validated data}
\item{"M"}{ - Manual data}
}}
}
\value{
a vector containing the date the station started reporting
and end reporting date
}
\description{
Dates supported by a station
}
\examples{
df <- get_station_dates(271, "M")
head(df)
}
