% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_and_dates.R
\name{get_station_parameters}
\alias{get_station_parameters}
\title{Parameters supported by a station}
\usage{
get_station_parameters(station_id, type = c("C", "V", "M"))
}
\arguments{
\item{station_id}{numeric code of the station}

\item{type}{The type of data to download. One of the following:
\itemize{
\item{"C"}{ - Crude data that has not been validated}
\item{"V"}{ - Validated data}
\item{"M"}{ - Manual data}
}}
}
\value{
a data.frame with the parameters supported by the station
}
\description{
Parameters supported by a station
}
\examples{
df <- get_station_parameters(271, "C")
head(df)
}
