% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Valid air quality parameters}
\format{A data frame with 30 rows and 1 variables:
\describe{
\item{parameter_code}{Abbreviation of the air quality parameter}
\item{parameter_name}{Name of the air quality parameter}
}}
\description{
Valid air quality parameters
}
\examples{
head(parameters)
}
