% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters_sinaica}
\alias{parameters_sinaica}
\title{Valid air quality parameters}
\format{A data frame with 55 rows and 2 variables:
\describe{
\item{parameter_code}{Abbreviation of the air quality parameter}
\item{parameter_name}{Name of the air quality parameter}
}}
\source{
\href{http://sinaica.inecc.gob.mx/}{SINAICA}
}
\description{
Valid air quality parameters
}
\examples{
head(parameters_sinaica)
}
