% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{params_sinaica}
\alias{params_sinaica}
\title{Valid air quality parameters}
\format{A data frame with 55 rows and 2 variables:
\describe{
\item{param_code}{Abbreviation of the air quality parameter}
\item{param_name}{Name of the air quality parameter}
}}
\source{
\href{https://sinaica.inecc.gob.mx/}{SINAICA}
}
\description{
Valid air quality parameters
}
\examples{
head(params_sinaica)
}
