% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinaica_param_data.R
\name{rsinaica-defunct}
\alias{rsinaica-defunct}
\alias{sinaica_param_data}
\title{Defunct Functions}
\usage{
sinaica_param_data(
  parameter,
  start_date,
  end_date,
  type = "Crude",
  remove_extremes = FALSE
)
}
\description{
These functions are Deprecated in this release of rsinaica, they will be
removed in a future version.
}
\keyword{internal}
