% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsinaica.R
\docType{package}
\name{rsinaica}
\alias{rsinaica}
\alias{rsinaica-package}
\title{\code{rsinaica} package}
\description{
Ready-made functions for downloading data from the Mexican
National Air Quality Information System (SINAICA). See
\url{http://sinaica.inecc.gob.mx} for more information.
}
\details{
See the README on
\href{https://github.com/diegovalle/rsinaica#readme}{GitHub}
}
