% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinaica_byparameter.R
\name{sinaica_byparameter}
\alias{sinaica_byparameter}
\title{Get data for all stations by parameter}
\usage{
sinaica_byparameter(parameter, start_date, end_date, autoclean = TRUE)
}
\arguments{
\item{parameter}{type of parameter to download
\itemize{
\item{"BEN"}{ - Benceno}
\item{"CH4"}{ - Metano}
\item{"CN"}{ - Carbono negro}
\item{"CO"}{ - Monóxido de carbono}
\item{"CO2"}{ - Dióxido de carbono}
\item{"DV"}{ - Dirección del viento}
\item{"H2S"}{ - Acido Sulfhídrico}
\item{"HCNM"}{ - Hidrocarburos no metánicos}
\item{"HCT"}{ - Hidrocarburos Totales}
\item{"HR"}{ - Humedad relativa}
\item{"HRI"}{ - Humedad relativa interior}
\item{"IUV"}{ - Índice de radiación ultravioleta}
\item{"NO"}{ - Óxido nítrico}
\item{"NO2"}{ - Dióxido de nitrógeno}
\item{"NOx"}{ - Óxidos de nitrógeno}
\item{"O3"}{ - Ozono}
\item{"PB"}{ - Presión Barométrica}
\item{"PM10"}{ - Partículas menores a 10 micras}
\item{"PM2.5"}{ - Partículas menores a 2.5 micras}
\item{"PP"}{ - Precipitación pluvial}
\item{"PST"}{ - Particulas Suspendidas totales}
\item{"RS"}{ - Radiación solar}
\item{"SO2"}{ - Dióxido de azufre}
\item{"TMP"}{ - Temperatura}
\item{"TMPI"}{ - Temperatura interior}
\item{"UVA"}{ - Radiación ultravioleta A}
\item{"VV"}{ - Radiación ultravioleta B}
\item{"XIL"}{ - Xileno}
}}

\item{start_date}{start date}

\item{end_date}{end date}

\item{autoclean}{wether to automatically remove invalid date and make sure
extreme values are turned to NAs}
}
\value{
data.frame with air quality data
}
\description{
Get data for all stations by parameter
}
\examples{
\dontrun{
## May take several seconds
df <- sinaica_byparameter("O3", "2015-10-14", "2015-10-14")
head(df)
}
}
