% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinaica_parameter.R
\name{sinaica_parameter}
\alias{sinaica_parameter}
\title{Get data for all stations by parameter}
\usage{
sinaica_parameter(parameter = c("BEN", "CH4", "CN", "CO", "CO2", "DV", "H2S",
  "HCNM", "HCT", "HR", "HRI", "IUV", "NO", "NO2", "NOx", "O3", "PB", "PM10",
  "PM2.5", "PP", "PST", "RS", "SO2", "TMP", "TMPI", "UVA", "UVB", "VV", "XIL"),
  start_date, end_date, autoclean = TRUE)
}
\arguments{
\item{parameter}{type of}

\item{start_date}{start date}

\item{end_date}{end date}

\item{autoclean}{clean the data automatically}
}
\value{
data.frame with air quality data
}
\description{
Get data for all stations by parameter
}
