% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinaica_station.R
\name{sinaica_station}
\alias{sinaica_station}
\title{Get data from a SINAICA reporting station}
\usage{
sinaica_station(station_id, parameter = c("BEN", "CH4", "CN", "CO", "CO2",
  "DV", "H2S", "HCNM", "HCT", "HR", "HRI", "IUV", "NO", "NO2", "NOx", "O3",
  "PB", "PM10", "PM2.5", "PP", "PST", "RS", "SO2", "TMP", "TMPI", "UVA", "UVB",
  "VV", "XIL"), type = c("C", "V", "M"), date, range = c("1", "2", "3",
  "4"))
}
\arguments{
\item{station_id}{See stations_sinaica}

\item{parameter}{type of parameter to download}

\item{type}{Crude, Validated or Manual data
Datos que se generan en las redes de monitoreo de la calidad del aire y
muestreo de contaminantes atmosféricos, que no han sido validados.}

\item{date}{day to download or start of range}

\item{range}{day, weekly, month}
}
\value{
data.frame with air quality data
}
\description{
Get data from a SINAICA reporting station
}
\examples{
stations_sinaica[which(stations_sinaica$station_name == "Xalostoc"), 1:5]
df <- sinaica_station(271, "O3", "C", "2015-09-11", "1")
head(df)

}
