% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters_and_dates.R
\name{get_station_dates}
\alias{get_station_dates}
\title{Dates supported by a station}
\usage{
get_station_dates(station_id, type = "Crude")
}
\arguments{
\item{station_id}{the numeric code correspongind to each station. See
\code{\link{stations_sinaica}} for a list of stations and their ids.}

\item{type}{The type of data to download. One of the following:
\itemize{
\item{"Crude"}{ - Crude data that has not been validated}
\item{"Validated"}{ - Validated data (may not be the most up-to-date)}
\item{"Manual"}{ - Manual data}
}}
}
\value{
a vector containing the date the station started reporting
and end reporting date
}
\description{
Dates supported by a station
}
\examples{
## id 271 is Xalostoc. See `stations_sinaica`
df <- get_station_dates(271, "Manual")
head(df)
}
