% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsinaica-package.R
\docType{package}
\name{rsinaica-package}
\alias{rsinaica}
\alias{rsinaica-package}
\title{Ready-made functions for downloading data from the Mexican
National Air Quality Information System (SINAICA). See
\url{https://sinaica.inecc.gob.mx} for more information.}
\description{
See the README on
\href{https://github.com/diegovalle/rsinaica#readme}{GitHub}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hoyodesmog.diegovalle.net/rsinaica/}
  \item \url{https://github.com/diegovalle/rsinaica}
  \item Report bugs at \url{https://github.com/diegovalle/rsinaica/issues}
}

}
\author{
\strong{Maintainer}: Diego Valle-Jones \email{diego@diegovalle.net}

}
\keyword{internal}
