% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinaica_bystation.R
\name{sinaica_bystation}
\alias{sinaica_bystation}
\title{Get air quaility data from a measuring station}
\usage{
sinaica_bystation(station_id, parameter, start_date, end_date, type = "Crude")
}
\arguments{
\item{station_id}{the numeric code correspongind to each station. See
\code{\link{stations_sinaica}} for a list of stations and their ids.}

\item{parameter}{type of parameter to download
\itemize{
\item{"BEN"}{ - Benceno}
\item{"CH4"}{ - Metano}
\item{"CN"}{ - Carbono negro}
\item{"CO"}{ - Monóxido de carbono}
\item{"CO2"}{ - Dióxido de carbono}
\item{"DV"}{ - Dirección del viento}
\item{"H2S"}{ - Acido Sulfhídrico}
\item{"HCNM"}{ - Hidrocarburos no metánicos}
\item{"HCT"}{ - Hidrocarburos Totales}
\item{"HR"}{ - Humedad relativa}
\item{"HRI"}{ - Humedad relativa interior}
\item{"IUV"}{ - Índice de radiación ultravioleta}
\item{"NO"}{ - Óxido nítrico}
\item{"NO2"}{ - Dióxido de nitrógeno}
\item{"NOx"}{ - Óxidos de nitrógeno}
\item{"O3"}{ - Ozono}
\item{"PB"}{ - Presión Barométrica}
\item{"PM10"}{ - Partículas menores a 10 micras}
\item{"PM2.5"}{ - Partículas menores a 2.5 micras}
\item{"PP"}{ - Precipitación pluvial}
\item{"PST"}{ - Particulas Suspendidas totales}
\item{"RS"}{ - Radiación solar}
\item{"SO2"}{ - Dióxido de azufre}
\item{"TMP"}{ - Temperatura}
\item{"TMPI"}{ - Temperatura interior}
\item{"UVA"}{ - Radiación ultravioleta A}
\item{"VV"}{ - Radiación ultravioleta B}
\item{"XIL"}{ - Xileno}
}}

\item{start_date}{start of range in YYYY-MM-DD format}

\item{end_date}{end of range from which to download data in YYYY-MM-DD format}

\item{type}{The type of data to download. One of the following:
\itemize{
\item{"Crude"}{ - Crude data that has not been validated}
\item{"Validated"}{ - Validated data (may not be the most up-to-date)}
\item{"Manual"}{ - Manually collected data that is sent to an external for lab analysis (may no be collected daily)}
}}
}
\value{
data.frame with air quality data
}
\description{
Download data from a single station by specifying a parameter and a date range
}
\examples{
stations_sinaica[which(stations_sinaica$station_name == "Xalostoc"), 1:5]
df <- sinaica_bystation(271, "O3", "2015-09-11", "2015-09-11", "Crude")
head(df)

}
\seealso{
Crude data comes from \href{http://sinaica.inecc.gob.mx/data.php},
validated data from \href{http://sinaica.inecc.gob.mx/data.php?tipo=V}, and
manual data from \href{http://sinaica.inecc.gob.mx/data.php?tipo=M}
}
