% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sinaica_station.R
\name{sinaica_station}
\alias{sinaica_station}
\title{Get data from a SINAICA reporting station}
\usage{
sinaica_station(station_id, parameter = c("BEN", "CH4", "CN", "CO", "CO2",
  "DV", "H2S", "HCNM", "HCT", "HR", "HRI", "IUV", "NO", "NO2", "NOx", "O3",
  "PB", "PM10", "PM2.5", "PM25", "PP", "PST", "RS", "SO2", "TMP", "TMPI", "UVA",
  "UVB", "VV", "XIL"), type = c("C", "V", "M"), date, range = c("1", "2",
  "3", "4"))
}
\arguments{
\item{station_id}{See stations_sinaica}

\item{parameter}{type of parameter to download}

\item{type}{The type of data to download. One of the following:
\itemize{
\item{"C"}{ - Crude data that has not been validated}
\item{"V"}{ - Validated data}
\item{"M"}{ - Manual data}
}}

\item{date}{day to download or start of range}

\item{range}{how much data to download
\itemize{
\item{"1"}{ - 1 day}
\item{"2"}{ - 1 week}
\item{"3"}{ - 2 weeks}
\item{"4"}{ - 1 month}
}}
}
\value{
data.frame with air quality data
}
\description{
Get data from a SINAICA reporting station
}
\examples{
stations_sinaica[which(stations_sinaica$station_name == "Xalostoc"), 1:5]
df <- sinaica_station(271, "O3", "C", "2015-09-11", "1")
head(df)

}
